Rewrote the spline interpolator
The spline interpolator is now written as a class. The constructor
takes the path to a file containing data and then builds all the
necessary vectors, matrices, and other parameters for the spline
interpolation to be done. Afterwards, the user can call the interpAt
member function to interpolate any amount of other values without
needlessly redoing any calculations.