Reorganized
Reorganized the project to allow flexibility with extending it in the
future. Interpolator classes now implement an interpolator interface,
so new interpolators can be added an used with any client code that only
depends on the interface. Interface classes for both 1 and 2 dimensional
interpolation are defined.

The original interpolation code has been moved into classes that inherit
the interpolator interface, so they serve as examples for adding new
interpolation methods.

Here is the list of commit messages for commits during the
reorganization (ordered from new to old):

    removed derivative and integral methods from the interface. still implemented in the base class.

    config and documentation updates.

    updated cmake config

    added data directory for unit tests.

    finished bilinear interpolator.

    Adding Bilinear interpolator.

    implemented 2D interpolator interface and base classes.

    added giblab-ci script.

    deleted version.h, it's generated by cmake.

    added setData function to the interpolation interface definition.

    removed old files.

    updated README for new interface.

    implemented spline interpolator.

    implemented base derivative and integration methods.

    implemented interpolator interface and base classes.