removed abstract base classes and all virtual functions
Removed all virtual functions in favor of std::function. One of the
original goals of this library was to allow runtime switching between
different interpolation algorithms. The "classic" way to achieve this
with inheritance and virtual functions. This requires creating a class
hierarchy that defines from a common base that defines the interface.

However, "Type Erasure" is an alternative method for achieving runtime
polymorphism that does not require inheritance. This is nice because you
can write classes that implement a given interface completely separate
from each other.

std::function is a new (C++11) class in the standard library that uses
type erasure to support binding to *any* callable at runtime. So, it is
no longer necessary to create a class hierarchy to support runtime
binding. We can write our interpolation classes completely independent
of each other if we want and use std::function for runtime switching.

As it turns out, inheritance is still useful for reusing code, so we are
still using it. But, we can now remove all virtual functions, because we
don't have to worry about needing to use a pointer-to-base-class to
achieve runtime bindings.