fix: performance bug in 1D interpolators
Found a bug in the 1D interpolators. The interpolation functions were
creating references to the interpolation data so that they could use
"value-style" method calls instead of "pointer-style" calls. So they had
something like this:

const VectorType& X = *(this->xView);
const VectorType& Y = *(this->yView);

The problem is that xView and yView are unique pointers of MapType, so I
*think* this was causing the data to be copied. Changing these to:

const MapType& X = *(this->xView);
const MapType& Y = *(this->yView);

dropped the runtime for interpolating a single point in a large data set
(200,000 points) from ~300 us to ~30 ns.